Add config, split to multiple files
